requirements.txt
Django==1.7b4
djangorestframework==2.3.13
